Web & API WIP

+++ WeekendPlanner_API/Controllers/ProtoEventController.cs
@@ -11,6 +11,7 @@ namespace WeekendPlanner_API.Controllers
     public class ProtoEventController : Controller
     {
         private readonly ProtoEventService protoEventService;
+        private const string sessionId = "sessionId";
 
         public ProtoEventController(ProtoEventService protoEventService)
         {
@@ -20,7 +21,7 @@ namespace WeekendPlanner_API.Controllers
         [HttpGet("allProtoEvents")]
         public async Task<List<Event>> Get()
         {
-            return await protoEventService.GetAsync();
+            return await protoEventService.GetAllAsync();
         }
 
         [HttpGet("oneProtoEvent")]
@@ -32,20 +33,56 @@ namespace WeekendPlanner_API.Controllers
         [HttpPost("createProtoEvent")]
         public async Task<IActionResult> createProtoEvent(Event newProtoEvent)
         {
-            await protoEventService.createProtoEvent(newProtoEvent);
-            return CreatedAtAction(nameof(Get), new { id = newProtoEvent.EventId }, newProtoEvent);
+            //Console.WriteLine(newProtoEvent.Time);
+            var userId = HttpContext.Session.GetString(sessionId);
+            Console.WriteLine(HttpContext.Session.GetString(sessionId));
+            Console.WriteLine(userId);
+            if (userId is not null)
+            {
+                if (newProtoEvent.CreatedBy == string.Empty)
+                {
+                    newProtoEvent.CreatedBy = userId;
+                }
+                await protoEventService.createProtoEvent(newProtoEvent);
+                return CreatedAtAction(nameof(Get), new { id = newProtoEvent.EventId }, newProtoEvent);
+            }
+            else
+            {
+                return Unauthorized(new { message = "Unauthorized" });
+            }
         }
 
         [HttpPut("updateProtoEvent")]
-        public async Task Put(string id, Event newProtoEvent)
+        public async Task<IActionResult> Put(string id, Event newProtoEvent)
         {
-            await protoEventService.updateProtoEvent(id, newProtoEvent);
+            var userId = HttpContext.Session.GetString(sessionId);
+            if (userId is not null) 
+            {
+                await protoEventService.updateProtoEvent(id, newProtoEvent);
+                return Ok(new { message = $"Event {id} was updated" });
+            }
+            else
+            {
+                return Unauthorized(new { message = "Unauthorized" });
+            }
+
         }
 
         [HttpDelete("deleteProtoEvent")]
-        public async Task Delete(string id)
+        public async Task<IActionResult> Delete(string id)
         {
-            await protoEventService.deleteProtoEvent(id);
+            var userId = HttpContext.Session.GetString(sessionId);
+
+            if (userId is not null) 
+            {
+                await protoEventService.deleteProtoEvent(id);
+                return Ok(new { message = $"Deleted Event {id}" });
+            }
+            else
+            {
+                return Unauthorized(new {message="Unauthorized"});
+            }
+
         }
 
     }
